extract date round by day function
