+++ src/stores/date.ts
@@ -1,14 +1,5 @@
 const zerofill = (s: number) => `0${s.toString()}`.slice(-2);
 
-// round by day
-const oneDayMs = 24 * 60 * 60 * 1000;
-export const dateFloor = (d: Date) => {
-  const offset = d.getTimezoneOffset() * 60 * 1000;
-  return new Date(
-    Math.floor((d.getTime() - offset) / oneDayMs) * oneDayMs + offset,
-  );
-};
-
 export const format = (date: number) => {
   const d = new Date(date);
   const y = d.getFullYear();
